+++ ValRougelike/Death/OnDeathChanges.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection.Emit;
 using UnityEngine;
+using static Deathlink.Common.DataObjects;
 
 namespace Deathlink.Death;
 
@@ -69,24 +70,17 @@ public static class OnDeathChanges
             }
         }
 
-        public static void TombstoneOnDeath(Player instance)
-        {
-            List<ItemDrop.ItemData> savedItems = new List<ItemDrop.ItemData>();
-            List<ItemDrop.ItemData> savedQuickslots = new List<ItemDrop.ItemData>();
+        public static void TombstoneOnDeath(Player instance) {
             List<ItemDrop.ItemData> playerItems = instance.m_inventory.GetAllItems();
-            List<ItemDrop.ItemData> playerNonSkillCheckItems = new List<ItemDrop.ItemData>();
             List<ItemDrop.ItemData> playerItemsWithoutNonSkillCheckedItems = new List<ItemDrop.ItemData>();
+            List<ItemDrop.ItemData> playerNonSkillCheckItems = new List<ItemDrop.ItemData>();
             Inventory inventory = instance.m_inventory;
 
-            List<ItemDrop.ItemData> playerItemsRemoved = new List<ItemDrop.ItemData>();
-            List<ItemDrop.ItemData> itemsToDrop = new List<ItemDrop.ItemData>();
+            Dictionary<ItemResults, List<ItemDrop.ItemData>> itemResults = new Dictionary<ItemResults, List<ItemDrop.ItemData>>();
 
-            GameObject tombstoneGo = null;
-            TombStone tombstone = null;
-
+            // Filter out the non-skill-checked items, this allows them to act differently than those that are skill checked
             string[] nonSkillCheckedItems = ValConfig.ItemsNotSkillChecked.Value.Split(',');
-            foreach (ItemDrop.ItemData item in playerItems)
-            {
+            foreach (ItemDrop.ItemData item in playerItems) {
                 if (item.m_dropPrefab != null && nonSkillCheckedItems.Contains(item.m_dropPrefab.name)) {
                     playerNonSkillCheckItems.Add(item);
                 } else {
@@ -94,38 +88,159 @@ public static class OnDeathChanges
                 }
             }
 
+            // Item destruction style
+            switch (Deathlink.pcfg().DeathStyle.itemLossStyle) {
+                case DataObjects.ItemLossStyle.None:
+                    Logger.LogDebug("No items destroyed on death.");
+                    itemResults.SafeInsertOrAppend(ItemResults.ItemSaved, playerItemsWithoutNonSkillCheckedItems);
+                    break;
+                case DataObjects.ItemLossStyle.DestroyNonWeaponArmor:
+                    Logger.LogDebug("Destroying non Equipment items.");
+                    itemResults.SafeInsertOrAppend(ItemResults.ItemLost, playerItemsWithoutNonSkillCheckedItems.GetNotEquipment());
+                    itemResults.SafeInsertOrAppend(ItemResults.EquipmentSaved, playerItemsWithoutNonSkillCheckedItems.GetEquipment());
+                    foreach (var item in playerItemsWithoutNonSkillCheckedItems.GetNotEquipment()) {
+                        instance.m_inventory.RemoveItem(item);
+                    }
+                    if (Deathlink.AzuEPILoaded) {
+                        Logger.LogDebug("AzuEPI| Removing non-equipment items from quickslots.");
+                        foreach (ItemDrop.ItemData item in AzuExtendedPlayerInventory.API.GetQuickSlotsItems()) {
+                            if (!item.IsEquipment()) {
+                                instance.m_inventory.RemoveItem(item);
+                                itemResults.SafeInsertOrAppend(ItemResults.ItemLost, item);
+                            } else {
+                                itemResults.SafeInsertOrAppend(ItemResults.EquipmentSaved, item);
+                            }
+                        }
+                    }
+                    break;
+                case DataObjects.ItemLossStyle.DestroyAll:
+                    Logger.LogDebug("Destroying all non skill checked items.");
+                    itemResults.SafeInsertOrAppend(ItemResults.ItemLost, playerItemsWithoutNonSkillCheckedItems);
+                    foreach(ItemDrop.ItemData item in playerItemsWithoutNonSkillCheckedItems) {
+                        instance.m_inventory.RemoveItem(item);
+                    }
+                    if (Deathlink.AzuEPILoaded) {
+                        Logger.LogDebug("AzuEPI| Destroying quickslot items.");
+                        itemResults.SafeInsertOrAppend(ItemResults.ItemLost, AzuExtendedPlayerInventory.API.GetQuickSlotsItems());
+                        AzuExtendedPlayerInventory.API.GetQuickSlotsItems().ForEach(item => instance.m_inventory.RemoveItem(item));
+                    }
+                    break;
+                case DataObjects.ItemLossStyle.DeathlinkBased:
+                    Logger.LogDebug("Destroying random items based on deathlink skill");
+                    itemResults = DetermineItemResultsByDeathlink(instance, playerItemsWithoutNonSkillCheckedItems);
+                    itemResults[ItemResults.EquipmentLost].ForEach(item => instance.m_inventory.RemoveItem(item));
+                    itemResults[ItemResults.ItemLost].ForEach(item => instance.m_inventory.RemoveItem(item));
+                    break;
+            }
+
+            // Determine what happens to the non-skill checked items, these items can also be added to the tombstone
+            if (Deathlink.pcfg().DeathStyle.nonSkillCheckedItemAction == NonSkillCheckedItemAction.Destroy) {
+                Logger.LogDebug($"Non skill checked being destroyed.");
+                itemResults.SafeInsertOrAppend(ItemResults.ItemLost, playerNonSkillCheckItems);
+            }
+            // Items not skill checked are requested to go to the tombstone
+            if (Deathlink.pcfg().DeathStyle.nonSkillCheckedItemAction == NonSkillCheckedItemAction.Save) {
+                Logger.LogDebug($"Non skill checked items being left on player.");
+            }
+
+            if (itemResults.ContainsKey(ItemResults.EquipmentSaved) && itemResults[ItemResults.EquipmentSaved].Count > 0 || itemResults.ContainsKey(ItemResults.ItemSaved) && itemResults[ItemResults.ItemSaved].Count > 0 || Deathlink.pcfg().DeathStyle.nonSkillCheckedItemAction == NonSkillCheckedItemAction.Tombstone && playerNonSkillCheckItems.Count() > 0) {
+                Logger.LogDebug($"Tombstone needed");
+                GameObject tombstoneGo = UnityEngine.Object.Instantiate<GameObject>(instance.m_tombstone, instance.GetCenterPoint(), instance.transform.rotation);
+                Inventory tombstoneInv = tombstoneGo.GetComponent<Container>().GetInventory();
+
+                // Save equipment and items to the tombstone, if configured to do so
+                if (Deathlink.pcfg().DeathStyle.itemSavedStyle == ItemSavedStyle.Tombstone) {
+                    Logger.LogDebug($"Adding saved items to tombstone.");
+                    if (itemResults.ContainsKey(ItemResults.EquipmentSaved) && itemResults[ItemResults.EquipmentSaved].Count > 0) {
+                        AddItemsToTombstone(tombstoneInv, itemResults[ItemResults.EquipmentSaved]);
+                        itemResults[ItemResults.EquipmentSaved].ForEach(item => instance.m_inventory.RemoveItem(item));
+                    }
+                    if (itemResults.ContainsKey(ItemResults.ItemSaved) && itemResults[ItemResults.ItemSaved].Count > 0) {
+                        AddItemsToTombstone(tombstoneInv, itemResults[ItemResults.ItemSaved]);
+                        itemResults[ItemResults.ItemSaved].ForEach(item => instance.m_inventory.RemoveItem(item));
+                    }
+                }
+
+                // Items not skill checked are requested to go to the tombstone
+                if (Deathlink.pcfg().DeathStyle.nonSkillCheckedItemAction == NonSkillCheckedItemAction.Tombstone) {
+                    Logger.LogDebug($"Adding saved non-skill checked items to tombstone.");
+                    AddItemsToTombstone(tombstoneInv, playerNonSkillCheckItems);
+                    playerNonSkillCheckItems.ForEach(item => instance.m_inventory.RemoveItem(item));
+                }
+
+                TombStone tombstone = tombstoneGo.GetComponent<TombStone>();
+                PlayerProfile playerProfile = Game.instance.GetPlayerProfile();
+                string name = playerProfile.GetName();
+                long playerId = playerProfile.GetPlayerID();
+                tombstone.Setup(name, playerId);
+                inventory.Changed();
+
+                // Whether or not to spawn the death marker on the map
+                if (ValConfig.ShowDeathMapMarker.Value) {
+                    Minimap.instance.AddPin(instance.transform.position, Minimap.PinType.Death, $"$hud_mapday {EnvMan.instance.GetDay(ZNet.instance.GetTimeSeconds())}", save: true, isChecked: false, 0L);
+                }
+            }
+        }
+
+        public static void AddItemsToTombstone(Inventory tombstone, List<ItemDrop.ItemData> transferItems) {
+            int size = Mathf.RoundToInt(Mathf.Sqrt(transferItems.Count())) + 1;
+            tombstone.m_width += size;
+            tombstone.m_height += size;
+            foreach (ItemDrop.ItemData item in transferItems) {
+                tombstone.m_inventory.Add(item);
+            }
+            tombstone.Changed();
+        }
+
+        internal static Dictionary<ItemResults, List<ItemDrop.ItemData>> DetermineItemResultsByDeathlink(Player player, List<ItemDrop.ItemData> playerItemsWithoutNonSkillCheckedItems) {
             float deathskillbonus = DeathProgressionSkill.DeathSkillCalculatePercentWithBonus();
             int items_to_keep = Mathf.RoundToInt(((Deathlink.pcfg().DeathStyle.maxItemsKept - Deathlink.pcfg().DeathStyle.minItemsKept) * deathskillbonus) + Deathlink.pcfg().DeathStyle.minItemsKept);
             int max_equipment_savable = Mathf.RoundToInt(((Deathlink.pcfg().DeathStyle.maxEquipmentKept - Deathlink.pcfg().DeathStyle.minEquipmentKept) * deathskillbonus) + Deathlink.pcfg().DeathStyle.minEquipmentKept);
             int numberOfItemsSavable = items_to_keep + max_equipment_savable;
 
-
-            // equipment specific
-            List<ItemDrop.ItemData> playerEquipment = Deathlink.shuffleList(playerItemsWithoutNonSkillCheckedItems.GetEquipment());
-            int max_eq_savable_by_type = max_equipment_savable;
-
-            if (Deathlink.AzuEPILoaded) {
-                foreach (ItemDrop.ItemData item in AzuExtendedPlayerInventory.API.GetQuickSlotsItems()) {
-                    if (nonSkillCheckedItems.Contains(item.m_shared.m_name)) {
-                        playerNonSkillCheckItems.Add(item);
-                    } else {
-                        playerItemsWithoutNonSkillCheckedItems.Add(item);
-                    }
-                }
+            Dictionary<ItemResults, List<ItemDrop.ItemData>> itemResults = new Dictionary<ItemResults, List<ItemDrop.ItemData>>(){
+                { ItemResults.EquipmentSaved, new List<ItemDrop.ItemData>() },
+                { ItemResults.EquipmentLost, new List<ItemDrop.ItemData>() },
+                { ItemResults.ItemSaved, new List<ItemDrop.ItemData>() },
+                { ItemResults.ItemLost, new List<ItemDrop.ItemData>() }
             }
+            ;
 
             // Equipment items are handled differently than resources etc
             int equipment_saved = 0;
-            foreach (var equipment in playerEquipment){
+            foreach (var equipment in Deathlink.shuffleList(playerItemsWithoutNonSkillCheckedItems.GetEquipment())) {
                 if (numberOfItemsSavable > 0) {
-                    if (RemoveEquipmentByStyle(equipment_saved, max_equipment_savable, instance, equipment, savedItems, numberOfItemsSavable, out numberOfItemsSavable, out equipment_saved)) {
+                    if (RemoveEquipmentByStyle(equipment_saved, max_equipment_savable, equipment, numberOfItemsSavable, out numberOfItemsSavable, out equipment_saved)) {
                         // If the item is not equipped but is still equipment, it should be saved since we have space for it
-                        savedItems.Add(equipment);
+                        Logger.LogDebug($"Save: {equipment.m_shared.m_name}");
+                        itemResults[ItemResults.EquipmentSaved].Add(equipment);
                     } else {
-                        playerItemsRemoved.Add(equipment);
+                        Logger.LogDebug($"Remove by style: {equipment.m_shared.m_name}");
+                        itemResults[ItemResults.EquipmentLost].Add(equipment);
                     }
                 } else {
-                    playerItemsRemoved.Add(equipment);
+                    Logger.LogDebug($"Save limit reached, removing: {equipment.m_shared.m_name}");
+                    itemResults[ItemResults.EquipmentLost].Add(equipment);
+                }
+            }
+
+            if (Deathlink.AzuEPILoaded) {
+                foreach (ItemDrop.ItemData item in AzuExtendedPlayerInventory.API.GetQuickSlotsItems()) {
+                    if (item.IsEquipment() && numberOfItemsSavable > 0) {
+                        numberOfItemsSavable -= 1;
+                        items_to_keep -= 1;
+                        itemResults[ItemResults.EquipmentSaved].Add(item);
+                        continue;
+                    } else {
+                        // Non equipment in a quickslot
+                        if (numberOfItemsSavable > 0) {
+                            numberOfItemsSavable -= 1;
+                            items_to_keep -= 1;
+                            itemResults[ItemResults.ItemSaved].Add(item);
+                        } else {
+                            itemResults[ItemResults.ItemLost].Add(item);
+                        }
+                    }
                 }
             }
 
@@ -135,102 +250,26 @@ public static class OnDeathChanges
                 // shuffle inventory items that are not equipment
                 foreach (var item in nonEquippableItems) {
                     if (numberOfItemsSavable > 0 && items_to_keep > 0) {
-                        Logger.LogDebug($"Saving {item.m_shared.m_name}");
-                        savedItems.Add(item);
+                        Logger.LogDebug($"NonEq| Saving: {item.m_shared.m_name}");
+                        itemResults[ItemResults.ItemSaved].Add(item);
                         numberOfItemsSavable -= 1;
                         items_to_keep -= 1;
                     } else {
-                        playerItemsRemoved.Add(item);
+                        Logger.LogDebug($"NonEq| Removing: {item.m_shared.m_name}");
+                        itemResults[ItemResults.ItemLost].Add(item);
                     }
                 }
             } else {
-                playerItemsRemoved.AddRange(nonEquippableItems);
-            }
-
-
-            // Quickslot changes happen after the inventory is cleaned up to avoid overstuffing the inventory
-            if (Deathlink.AzuEPILoaded)
-            {
-                Logger.LogDebug($"Quickslot items found {AzuExtendedPlayerInventory.API.GetQuickSlotsItems().Count}");
-                // If the item was not saved, it should be removed from the quickslots
-                foreach (var item in AzuExtendedPlayerInventory.API.GetQuickSlotsItems())
-                {
-                    if (savedItems.Contains(item)) { continue; }
-                    Logger.LogDebug($"Removing quickslot item that was not saved {item.m_shared.m_name}");
-                    playerItemsRemoved.Add(item);
-                    //instance.UnequipItem(item);
-                    //instance.m_inventory.RemoveItem(item);
+                Logger.LogDebug($"NonEq| Removing: {nonEquippableItems.Count} due to no saves remaining.");
+                foreach(var item in nonEquippableItems) {
+                    itemResults[ItemResults.ItemLost].Add(item);
                 }
             }
 
-            // Handle items that are defined in the non-skill-checked section
-            // we do this right after clearing the inventory to allow creating a tombstone from the empty inventory (which can just contain our non-skill handled items)
-            switch (Deathlink.pcfg().DeathStyle.nonSkillCheckedItemAction) {
-                case DataObjects.NonSkillCheckedItemAction.Tombstone:
-                    Logger.LogDebug($"Dropping non-skill-checked items on death ({playerNonSkillCheckItems.Count()})");
-                    itemsToDrop.AddRange(playerNonSkillCheckItems);
-                    break;
-                case DataObjects.NonSkillCheckedItemAction.Destroy:
-                    Logger.LogDebug($"Destroying non-skill-checked items on death ({playerNonSkillCheckItems.Count()})");
-                    playerItemsRemoved.AddRange(playerNonSkillCheckItems);
-                    break;
-                case DataObjects.NonSkillCheckedItemAction.Save:
-                    Logger.LogDebug($"Saving non-skill-checked items on death ({playerNonSkillCheckItems.Count()})");
-                    savedItems.AddRange(playerNonSkillCheckItems);
-                    break;
-            }
-
-            bool tombstoneCreated = false;
-            if (Deathlink.pcfg().DeathStyle.itemLossStyle == DataObjects.ItemLossStyle.None && playerItemsRemoved.Count > 0) {
-                Logger.LogDebug($"Items failed skillcheck {playerItemsRemoved.Count} items on death, dropping to tombestone.");
-                itemsToDrop.AddRange(playerItemsRemoved);
-            }
-
-            // Empty the inventory, we already have everything that is getting saved copied off.
-            foreach (var item in playerItemsRemoved) {
-                instance.m_inventory.RemoveItem(item);
-            }
-
-            if (Deathlink.pcfg().DeathStyle.itemSavedStyle == DataObjects.ItemSavedStyle.Tombstone && savedItems.Count > 0) {
-                Logger.LogDebug($"Saving {savedItems.Count} Items to tombstone");
-                itemsToDrop.AddRange(savedItems);
-            }
-
-            if (itemsToDrop.Count > 0) {
-                Logger.LogDebug("Saving droppable items to tombstone");
-
-                tombstoneCreated = true;
-                tombstoneGo = UnityEngine.Object.Instantiate<GameObject>(instance.m_tombstone, instance.GetCenterPoint(), instance.transform.rotation);
-                AddItemsToTombstone(tombstoneGo.GetComponent<Container>().GetInventory(), itemsToDrop);
-                foreach (var item in itemsToDrop) {
-                    instance.m_inventory.RemoveItem(item);
-                }
-                // gameObject.GetComponent<Container>().GetInventory().MoveInventoryToGrave(instance.m_inventory);
-                tombstone = tombstoneGo.GetComponent<TombStone>();
-                PlayerProfile playerProfile = Game.instance.GetPlayerProfile();
-                string name = playerProfile.GetName();
-                long playerId = playerProfile.GetPlayerID();
-                tombstone.Setup(name, playerId);
-                inventory.Changed();
-            }
-
-            // Whether or not to spawn the death marker on the map
-            if (ValConfig.ShowDeathMapMarker.Value && tombstoneCreated) {
-                Minimap.instance.AddPin(instance.transform.position, Minimap.PinType.Death, $"$hud_mapday {EnvMan.instance.GetDay(ZNet.instance.GetTimeSeconds())}", save: true, isChecked: false, 0L);
-            }
+            return itemResults;
         }
 
-        public static void AddItemsToTombstone(Inventory tombstone, List<ItemDrop.ItemData> transferItems) {
-            int size = Mathf.RoundToInt(Mathf.Sqrt(transferItems.Count())) + 1;
-            tombstone.m_width = size;
-            tombstone.m_height = size;
-            foreach (ItemDrop.ItemData item in transferItems) {
-                tombstone.m_inventory.Add(item);
-            }
-            tombstone.Changed();
-        }
-
-        internal static bool RemoveEquipmentByStyle(int equipment_saved, int max_equipment_savable, Player instance, ItemDrop.ItemData equipment, List<ItemDrop.ItemData> saved_equipment, int numberOfItemsSavable, out int remainingsaves, out int equipment_saved_count)
+        internal static bool RemoveEquipmentByStyle(int equipment_saved, int max_equipment_savable, ItemDrop.ItemData equipment, int numberOfItemsSavable, out int remainingsaves, out int equipment_saved_count)
         {
             equipment_saved_count = 0;
             remainingsaves = 0;